+++ src/routes/Auth.js
@@ -1,31 +1,8 @@
+import AuthForm from 'components/AuthForm'
 import { authService, firebaseInstance } from 'fbase'
-import React, { useState } from 'react'
+import React from 'react'
 
 const Auth = () => {
-  const [email, setEmail] = useState('')
-  const [password, setPassword] = useState('')
-  const [newAccount, setNewAccount] = useState(true)
-  const [error, setError] = useState('')
-
-  const onChange = (event) => {
-    const { target: { name, value } } = event // event.target. ...
-    if (name === 'email') setEmail(value)
-    else if (name === 'password') setPassword(value)
-  }
-  const onSubmit = async(event) => {
-    event.preventDefault()
-    try {
-      let data
-      if (newAccount) data = await authService.createUserWithEmailAndPassword(
-          email, password
-        )
-      else data = await authService.signInWithEmailAndPassword(email, password)
-      console.log('submit!', data)
-    } catch (error) {
-      setError(error.message)
-    }
-  }
-  const toggleAccount = () => setNewAccount(prev => !prev)
   const onSocialClick = async (event) => {
     const {
       target : { name },
@@ -43,32 +20,7 @@ const Auth = () => {
   }
   return (
     <div>
-      <form onSubmit={onSubmit}>
-        <input
-          name="email"
-          type="email"
-          placeholder="Email"
-          required
-          value={email}
-          onChange={onChange}
-        />
-        <input
-          name="password"
-          type="password"
-          placeholder="Password"
-          required
-          value={password}
-          onChange={onChange}
-        />
-        <input
-          type="submit"
-          value={ newAccount ? 'Create Account' : 'Sign In' }
-        />
-        {error}
-      </form>
-      <span onClick={toggleAccount}>
-        {newAccount ? 'Sign in' : 'Create Account'}
-      </span>
+      <AuthForm />
       <div>
         <button 
           name="google"
